+++ MyLibrary/.swiftpm/xcode/xcshareddata/xcschemes/MyLibrary.xcscheme
@@ -48,6 +48,26 @@
                ReferencedContainer = "container:">
             </BuildableReference>
          </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "Integration Tests"
+               BuildableName = "Integration Tests"
+               BlueprintName = "Integration Tests"
+               ReferencedContainer = "container:">
+            </BuildableReference>
+         </TestableReference>
+         <TestableReference
+            skipped = "NO">
+            <BuildableReference
+               BuildableIdentifier = "primary"
+               BlueprintIdentifier = "Unit Tests"
+               BuildableName = "Unit Tests"
+               BlueprintName = "Unit Tests"
+               ReferencedContainer = "container:">
+            </BuildableReference>
+         </TestableReference>
       </Testables>
    </TestAction>
    <LaunchAction
